# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DATN API",
        "description": "API for DATN project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/supply-logs": {
            "get": {
                "tags": [
                    "Supply Logs"
                ],
                "summary": "Danh sách lịch sử vật tư",
                "description": "Lấy danh sách tất cả lịch sử vật tư (nhập, xuất, điều chỉnh) với hỗ trợ lọc",
                "operationId": "getSupplyLogs",
                "parameters": [
                    {
                        "name": "action_type",
                        "in": "query",
                        "description": "Lọc theo loại hành động (in, out, adjustment)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "supply_id",
                        "in": "query",
                        "description": "Lọc theo ID vật tư",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "performed_by",
                        "in": "query",
                        "description": "Lọc theo người thực hiện",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Lọc từ ngày (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Lọc đến ngày (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang (20 kết quả/trang)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách lịch sử vật tư",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supply-logs/{id}": {
            "get": {
                "tags": [
                    "Supply Logs"
                ],
                "summary": "Chi tiết lịch sử vật tư",
                "description": "Lấy thông tin chi tiết của một bản ghi lịch sử vật tư",
                "operationId": "getSupplyLog",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết lịch sử vật tư",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supply-logs/supply/{supplyId}": {
            "get": {
                "tags": [
                    "Supply Logs"
                ],
                "summary": "Lịch sử vật tư theo ID",
                "description": "Lấy toàn bộ lịch sử nhập/xuất/điều chỉnh của một vật tư",
                "operationId": "getSupplyHistoryLogs",
                "parameters": [
                    {
                        "name": "supplyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lịch sử vật tư",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supply-logs/activities/recent": {
            "get": {
                "tags": [
                    "Supply Logs"
                ],
                "summary": "Hoạt động vật tư gần đây",
                "description": "Lấy danh sách hoạt động vật tư gần đây nhất",
                "operationId": "getRecentActivities",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số bản ghi tối đa (mặc định 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hoạt động gần đây",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/supply-logs/summary/movement": {
            "get": {
                "tags": [
                    "Supply Logs"
                ],
                "summary": "Tóm tắt vận động vật tư",
                "description": "Lấy tóm tắt vận động vật tư (nhập, xuất, điều chỉnh) theo khoảng thời gian",
                "operationId": "getMovementSummary",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Từ ngày (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Đến ngày (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tóm tắt vận động",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_inbound": {
                                                    "type": "integer"
                                                },
                                                "total_outbound": {
                                                    "type": "integer"
                                                },
                                                "total_adjustments": {
                                                    "type": "integer"
                                                },
                                                "total_transfers": {
                                                    "type": "integer"
                                                },
                                                "inbound_value": {
                                                    "type": "number"
                                                },
                                                "outbound_value": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based based security",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}